update isValid method and update the handlemessage
isValid: last version queen can move to the place with arrow. Add condition to avoid that happened.

handlemessage: last version black and white board are not same. Now it is same. But when black play the chess, white did nothing. I think the code did not check which color we played. Default is black, so first move is good. then did not change when we play white, so the white did nothing